requirements.txt
fastapi
requests
motor
httpx
python-dotenv